Refactor: Moved expression parser files into `expr/atomic/literals` and `expr/atomic/operators` to improve navigation, scalability, and reduce cognitive complexity.
> The Lord had said to Abram, "Go from your country, your people and your father's household to the land I will show you." – Genesis 12:1
