+++ src/jesus/parser/grammar/expr/atomic/literals/number_rule.hpp
@@ -1,17 +1,16 @@
-
 #pragma once
-#include "../grammar_rule.hpp"
+#include "../../../grammar_rule.hpp"
 
 /**
- * @brief Matches a string literal.
+ * @brief Matches a number literal.
  */
-class StringRule : public IGrammarRule
+class NumberRule : public IGrammarRule
 {
 public:
     std::unique_ptr<Expr> parse(ParserContext &ctx) override;
 
     std::string toStr(GrammarRuleHashTable &visitedTable) const override
     {
-        return "String";
+        return "Number";
     }
 };
