+++ src/jesus/parser/grammar/expr/atomic/operators/comparison_rule.cpp
@@ -1,5 +1,5 @@
 #include "comparison_rule.hpp"
-#include "../../../ast/expr/binary_expr.hpp"
+#include "../../../../../ast/expr/binary_expr.hpp"
 
 std::unique_ptr<Expr> ComparisonRule::parse(ParserContext &ctx)
 {
