+++ src/jesus/parser/grammar/expr/atomic/operators/equality_rule.cpp
@@ -1,5 +1,5 @@
 #include "equality_rule.hpp"
-#include "../../../ast/expr/binary_expr.hpp"
+#include "../../../../../ast/expr/binary_expr.hpp"
 
 std::unique_ptr<Expr> EqualityRule::parse(ParserContext &ctx)
 {
