+++ src/jesus/parser/grammar/expr/atomic/operators/multiplication_rule.cpp
@@ -1,5 +1,5 @@
 #include "multiplication_rule.hpp"
-#include "../../../ast/expr/binary_expr.hpp"
+#include "../../../../../ast/expr/binary_expr.hpp"
 
 std::unique_ptr<Expr> MultiplicationRule::parse(ParserContext &ctx)
 {
