+++ src/jesus/parser/grammar/expr/atomic/operators/multiplication_rule.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../grammar_rule.hpp"
+#include "../../../grammar_rule.hpp"
 
 class MultiplicationRule : public IGrammarRule
 {
